+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.10.0
+
+### New features
+
+- Add extension functions for starting a parse. Instead of writing
+  e.g. `parse(p, input)` you can write `p.parse(input)`.
+
 ## 0.9.0
 
 ### New features